fix(log): Fix dev config write to bytes logger.
Under the dev config, where `sys.isatty()` is true, the console renderer
used emits `str` but was writing to `BytesLogger` which caused an error
when the bytes logger would try to concat a byte string to the `str`.
So we don't use the `BytesLogger` for dev.

Also, removed the multiple traceback formatters from the prod
configuration.