feat(service)!: removes the service authorization methods.
I don't want to preempt the requirements here until I get the user
model integrated.

There are some specific requirements that I will have for being able
to control the attributes that may be modified by a user depending on
their type and the model, so I don't want to pre-suppose anything about
that implementation.