feat!(service): Adds `service.RepositoryService` type.
The coupling between the base service object and a repository type was
too inflexible to be useful when building an app that doesn't need a
repository.

`service.Service` is no longer aware of `repository_type` and has no
default implementations defined.

To get the same behavior as previous versions, service objects should
now inherit from `service.RepositoryService`.

This is for #170, but #170 is broader than this change. We still need
to make the repository service not _need_ to use SQLAlchemy ORM types
as the model type.