fix(reloading): use `uvicorn.run()`
Logic for reloading is in `uvicorn.run()`, but to use it we need to let
uvicorn manage the event loop for us.

So this commit moves the database/cache readiness checks into a
starlite, "before_startup" hook handler, so that they are conducted
after uvicorn has setup the event loop.