refactor: use constants for test/local env checks.
We are starting to amass a bit of functionality that is condition upon
the environment setting. E.g., sentry init disabled by default in
test and local environments, alternate logging config if environment is
local, and server reloading automatically configured for local
environment. So, it makes sense to have a bit more structure around
these checks.

Closes #244